// frontend/frontendERP/src/app/api/models/vente.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LigneRequest } from '../models/ligne-request';
import { VenteType } from '../models/vente-type';
export interface Vente {
  dateVente?: string;
  id: string;
  lignes?: Array<LigneRequest>;
  montantTotal: number;
  type?: VenteType;
  vendeurId?: string;
}